+++ src/DurableTask.SqlServer.AzureFunctions/SqlScaleMonitor.cs
@@ -19,15 +19,21 @@ namespace DurableTask.SqlServer.AzureFunctions
         static readonly ScaleStatus NoScaleVote = new ScaleStatus { Vote = ScaleVote.None };
         static readonly ScaleStatus ScaleOutVote = new ScaleStatus { Vote = ScaleVote.ScaleOut };
 
-        readonly SqlOrchestrationService service;
         readonly SqlMetricsProvider metricsProvider;
 
         int? previousWorkerCount = -1;
 
-        public SqlScaleMonitor(SqlOrchestrationService service, string taskHubName, SqlMetricsProvider sqlMetricsProvider)
+        public SqlScaleMonitor(string taskHubName, SqlMetricsProvider sqlMetricsProvider)
         {
-            this.service = service ?? throw new ArgumentNullException(nameof(service));
-            this.Descriptor = new ScaleMonitorDescriptor($"DurableTask-SqlServer:{taskHubName ?? "default"}");
+            // Scalers in Durable Functions are shared for all functions in the same task hub.
+            // So instead of using a function ID, we use the task hub name as the basis for the descriptor ID.
+            string id = $"DurableTask-SqlServer:{taskHubName ?? "default"}";
+
+#if FUNCTIONS_V4
+            this.Descriptor = new ScaleMonitorDescriptor(id: id, functionId: id);
+#else
+            this.Descriptor = new ScaleMonitorDescriptor(id);
+#endif
             this.metricsProvider = sqlMetricsProvider ?? throw new ArgumentNullException(nameof(sqlMetricsProvider));
         }