+++ src/DurableTask.SqlServer.AzureFunctions/SqlTargetScaler.cs
@@ -1,22 +1,24 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+#if FUNCTIONS_V4
 namespace DurableTask.SqlServer.AzureFunctions
 {
     using System.Threading.Tasks;
     using Microsoft.Azure.WebJobs.Host.Scale;
 
-#if !NETSTANDARD
     public class SqlTargetScaler : ITargetScaler
     {
         readonly SqlMetricsProvider sqlMetricsProvider;
-        readonly TargetScalerResult scaleResult;
 
-        public SqlTargetScaler(string functionId, SqlMetricsProvider sqlMetricsProvider)
+        public SqlTargetScaler(string taskHubName, SqlMetricsProvider sqlMetricsProvider)
         {
             this.sqlMetricsProvider = sqlMetricsProvider;
-            this.scaleResult = new TargetScalerResult();
-            this.TargetScalerDescriptor = new TargetScalerDescriptor(functionId);
+
+            // Scalers in Durable Functions are shared for all functions in the same task hub.
+            // So instead of using a function ID, we use the task hub name as the basis for the descriptor ID.
+            string id = $"DurableTask-SqlServer:{taskHubName ?? "default"}";
+            this.TargetScalerDescriptor = new TargetScalerDescriptor(id);
         }
 
         public TargetScalerDescriptor TargetScalerDescriptor { get; private set; }
@@ -24,9 +26,11 @@ namespace DurableTask.SqlServer.AzureFunctions
         public async Task<TargetScalerResult> GetScaleResultAsync(TargetScalerContext context)
         {
             SqlScaleMetric sqlScaleMetric = await this.sqlMetricsProvider.GetMetricsAsync();
-            this.scaleResult.TargetWorkerCount = sqlScaleMetric.RecommendedReplicaCount;
-            return this.scaleResult;
+            return new TargetScalerResult
+            {
+                TargetWorkerCount = sqlScaleMetric.RecommendedReplicaCount,
+            };
         }
     }
-#endif
 }
+#endif